if checks after each polling while loop--if cmdlet fails then write output saying so and exit script

+++ Invoke-ContainerToDataFactory.ps1
@@ -13,12 +13,6 @@ Param(
     [string] $PipelineName
 )
 
-
-# ./Invoke-ContainerToAdfLocal.ps1 -ContainerImage jarvisrob/scrapereiv -ContainerResourceGroup rob-real-estate-test -ContainerName scrapereiv -ContainerVolumeMountPath /out -StorageAccountResourceGroup rob-re-store -StorageAccountName robrestore -StorageAccountKey $StorageAccountKey -FileShareName scrape -BlobContainerName adf-test-blob -DataFactoryResourceGroup rob-re-adf -DataFactoryName readftest -PipelineName CopyPipeline-RealEstate
-
-# Connection to Azure account
-# Login-AzureRmAccount
-
 # Naming the file produced
 $y = Get-Date -UFormat %Y
 $m = Get-Date -UFormat %m
@@ -39,13 +33,18 @@ While (-not (($ContainerInfo.State -eq "Succeeded") -or ($ContainerInfo.State -e
     Start-Sleep -Seconds 10
     $ContainerInfo = Get-AzureRmContainerGroup -ResourceGroupName $ContainerResourceGroup -Name $ContainerName
 }
-Write-Output "Container finished. Final instance log:"
-Get-AzureRmContainerInstanceLog -ResourceGroupName $ContainerResourceGroup -ContainerGroupName $ContainerName -Tail 20
-Write-Output "Killing container"
-Remove-AzureRmContainerGroup -ResourceGroupName $ContainerResourceGroup -Name $ContainerName
+If ($ContainerInfo.State -eq "Succeeded") {
+    Write-Output "Container finished. Final instance log:"
+    Get-AzureRmContainerInstanceLog -ResourceGroupName $ContainerResourceGroup -ContainerGroupName $ContainerName -Tail 20
+    Write-Output "Killing container"
+    Remove-AzureRmContainerGroup -ResourceGroupName $ContainerResourceGroup -Name $ContainerName
+} Else {
+    Write-Output "ERROR: Container failed. Exiting script."
+    Exit
+}
 
 # Copy file(s) from file storage to blob
-Write-Output "Copying file(s) to staging blob ..."
+Write-Output "Copying file(s) to staging blob"
 Set-AzureRmCurrentStorageAccount -ResourceGroupName $StorageAccountResourceGroup -Name  $StorageAccountName
 Start-AzureStorageBlobCopy -SrcShareName $FileShareName -SrcFilePath "/$OutFileName" -DestContainer $BlobContainerName
 $BlobCopyState = Get-AzureStorageBlobCopyState -Blob $OutFileName -Container $BlobContainerName
@@ -54,7 +53,12 @@ While (-not (($BlobCopyState.Status -eq "Success") -or ($BlobCopyState.Status -e
     Start-Sleep -Seconds 2
     $BlobCopyState = Get-AzureStorageBlobCopyState -Blob $OutFileName -Container $BlobContainerName
 }
-Write-Output "Copy complete"
+If ($BlobCopyState.Status -eq "Success") {
+    Write-Output "Copy complete"
+} Else {
+    Write-Output "ERROR: Copy failed. Exiting script."
+    Exit
+}
 
 # Invoke the Data Factory pipeline
 Write-Output "Invoking the Data Factory pipeline"
@@ -67,7 +71,11 @@ While (-not (($PipelineRunInfo.Status -eq "Succeeded") -or ($PipelineRunInfo.Sta
     Start-Sleep -Seconds 10
     $PipelineRunInfo = Get-AzureRmDataFactoryV2PipelineRun -ResourceGroupName $DataFactoryResourceGroup -DataFactoryName $DataFactoryName -PipelineRunId $PipelineRunId
 }
-Write-Output "Pipline run complete"
-
+If ($PipelineRunInfo.Status -eq "Succeeded") {
+    Write-Output "Pipline run complete"
+} Else {
+    Write-Output "ERROR: Pipeline run failed. Exiting script."
+    Exit
+}
 
 Write-Output "Script completed"
